📁 refactor[UserConfirmationScreen]: restructure to Split UI and Create Hooks
